ui: Move discovery chain component
