+++ ui/packages/consul-ui/app/components/consul/discovery-chain/index.scss
@@ -1,4 +1,5 @@
-@import './discovery-chain/index';
+@import './skin';
+@import './layout';
 .discovery-chain {
   @extend %discovery-chain;
 }
