+++ ui/packages/consul-ui/app/components/consul/discovery-chain/route-card/index.hbs
@@ -1,28 +1,32 @@
-<a class="route-card" onclick={{onclick}} id={{item.ID}}>
-  <header class={{if (eq path.value '/') 'short'}}>
-{{#if (gt item.Definition.Match.HTTP.Methods.length 0) }}
+<a
+  class="route-card"
+  onclick={{@onclick}}
+  id={{@item.ID}}
+>
+  <header class={{if (eq this.path.value '/') 'short'}}>
+{{#if (gt @item.Definition.Match.HTTP.Methods.length 0) }}
     <ul class="match-methods">
-      {{#each item.Definition.Match.HTTP.Methods as |item|}}
+      {{#each @item.Definition.Match.HTTP.Methods as |item|}}
         <li>{{item}}</li>
       {{/each}}
     </ul>
 {{/if}}
     <dl>
       <dt>
-      {{path.type}}
+        {{this.path.type}}
       </dt>
       <dd>
-        {{path.value}}
+        {{this.path.value}}
       </dd>
     </dl>
   </header>
-{{#if (gt item.Definition.Match.HTTP.Header.length 0) }}
+{{#if (gt @item.Definition.Match.HTTP.Header.length 0) }}
   <section class="match-headers">
     <header data-tooltip="Header">
       <h4>Headers</h4>
     </header>
     <dl>
-  {{#each item.Definition.Match.HTTP.Header as |item|}}
+  {{#each @item.Definition.Match.HTTP.Header as |item|}}
       <dt>
         {{item.Name}}
       </dt>
@@ -33,13 +37,13 @@
     </dl>
   </section>
 {{/if}}
-{{#if (gt item.Definition.Match.HTTP.QueryParam.length 0) }}
+{{#if (gt @item.Definition.Match.HTTP.QueryParam.length 0) }}
   <section class="match-queryparams">
     <header data-tooltip="Query Params">
       <h4>Query Params</h4>
     </header>
     <dl>
-  {{#each item.Definition.Match.HTTP.QueryParam as |item|}}
+  {{#each @item.Definition.Match.HTTP.QueryParam as |item|}}
       <dt>
         {{item.Name}}
       </dt>
