+++ ui/packages/consul-ui/app/components/consul/discovery-chain/splitter-card/index.hbs
@@ -1,0 +1,13 @@
+<div
+  ...attributes
+>
+  <a
+    id={{@item.ID}}
+    class="splitter-card"
+    onclick={{optional @onclick}}
+  >
+    <header>
+      <h3>{{@item.Name}}</h3>
+    </header>
+  </a>
+</div>
