+++ ui/packages/consul-ui/app/components/consul/discovery-chain/utils.js
