--- ui/packages/consul-ui/app/components/resolver-card/index.js
@@ -1,5 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({
-  tagName: '',
-});
