--- ui/packages/consul-ui/app/components/splitter-card/index.hbs
@@ -1,5 +1,0 @@
-<a class="splitter-card" onclick={{onclick}} id={{item.ID}}>
-  <header>
-    <h3>{{item.Name}}</h3>
-  </header>
-</a>
