--- ui/packages/consul-ui/app/components/splitter-card/index.js
@@ -1,3 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({});
