+++ ui/packages/consul-ui/app/styles/components.scss
@@ -29,7 +29,6 @@
 @import './components/auth-form';
 @import './components/auth-modal';
 @import './components/oidc-select';
-@import './components/discovery-chain';
 @import './components/empty-state';
 @import './components/tabular-details';
 @import './components/tabular-collection';
@@ -56,6 +55,7 @@
 @import 'consul-ui/components/notice';
 @import 'consul-ui/components/modal-dialog';
 
+@import 'consul-ui/components/consul/discovery-chain';
 @import 'consul-ui/components/consul/exposed-path/list';
 @import 'consul-ui/components/consul/external-source';
 @import 'consul-ui/components/consul/kind';
