--- ui/packages/consul-ui/app/styles/components/discovery-chain/index.scss
@@ -1,2 +1,0 @@
-@import './skin';
-@import './layout';
