+++ ui/packages/consul-ui/app/templates/dc/services/show/routing.hbs
@@ -1,6 +1,8 @@
 <div id="routing" class="tab-section">
   <div role="tabpanel">
-    <DiscoveryChain @chain={{chain.Chain}} />
+    <Consul::DiscoveryChain
+      @chain={{chain.Chain}}
+    />
   </div>
 </div>
